// src/masters/inspection36/dto/create-inspection36.dto.ts
export class CreateInspection36Dto {
  address?: string;
  city?: string;
  stzip?: string;
  date?: string;
  fileno?: string;
  appraiser?: string;
  borrower?: string;
  timein?: string;
  timeout?: string;
  proptype?: string;
  dwelling_style?: string;
  attachment_type?: string;
  status?: string;

  arrive?: any;
  curb?: any;
  exterior?: any;
  yard?: any;
  outbuildings?: any;
  mainlevel?: any;
  upperlevel?: any;
  belowgrade?: any;
  adu?: any;
  final?: any;
}